Remove use of Categories from DockWindowTheme.
